Mise à jour de la question1

+++ TP2/Question1/allocateur.h
@@ -2,16 +2,17 @@
 #define ALLOCATEUR_H
 
 #include <stddef.h>
+#include <stdint.h>
 
 typedef struct HEADER_TAG {
 	struct HEADER_TAG *ptr_next;
 	size_t bloc_size;
-	long magic_number;
+	uint64_t magic_number;
 } HEADER;
 
 
 void init_heap();
-void *malloc_3is(int size);
+void *malloc_3is(size_t size);
 void free_3is(void *ptr);
 
 #endif
